# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "routes::user"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizeResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "InternalErrorResponse": {
        "description": "An internal error occurred",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
...
      "UnauthorizeResponse": {
        "description": "User not authorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status_code",
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "status_code": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          }
        }
      },
      "UserResponse": {
        "description": "",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ]
}